server-javascript: prevent G-code commands being spammed after unpause
After a print is unpaused, there's a possibility that
a printer processed multiple commands. Therefore, it
likely sent multiple "ok's" which will cause the system
to send a new G-code command to the printer for each
"ok" which could overwhelm the printer. This would result
in those commands being ignored (sometimes almost silently).

Also, a redundant type check was removed